Change "Handler" code to only be called once
The class which mutates the message from the on-wire
format to the format that's usable by the user (relabels fields,
filters fields, etc...) was being called eached time a new handler
was being invoked. Instead, this changes it so that it is only
invoked once, and then that relabeled message is passed to
each subsequent callback function.

This also simplifies the handler to just be a function
rather than a full blown class. Albeit that function
is a lambda over the fields that were previously passed
to the class.

This also adds an enum for the full list of services that
TD: Ameritrade offers. This enum can then be used to initialize
things like the internal handlers list.