Continue to pass by copy, not by deference
umhelp pointed out quite clearly that the old behaviour of the API
was that each callback handler was called with a copy of the object
and not the object itself. This adds a new wrapper function, and
option keyword argument to indicate whether or not you want
a copy to be passed back.